# Feature/performance rework