--- a/lib/mod.ts
+++ b/lib/mod.ts
@@ -1,4 +1,4 @@
-import { Lazy } from './iterators.ts';
+import { Lazy } from './lazy.ts';
 
 export { Lazy };