--- a/test/deps/std/testing/asserts.ts
+++ b/test/deps/std/testing/asserts.ts
@@ -1,0 +1,1 @@
+export * from 'https://deno.land/std@v0.3.2/testing/asserts.ts';