Benchmarks from before (using the new benchmarks script):

```
=== node ===
Map benchmark: 0.3344795929312706ms
Map for-of benchmark: 0.21867035821676253ms
Select benchmarks: 4.502886601561308ms
Select for-of benchmark: 3.4375009231626987ms
=== deno ===
Map benchmark: 1.0657ms
Map for-of benchmark: 0.2174ms
Select benchmarks: 4.2926ms
Select for-of benchmark: 3.2965ms
```

Benchmarks after the rework:

```
=== node ===
Map benchmark: 0.34207717384696007ms
Map for-of benchmark: 0.22262977133989334ms
Select benchmarks: 1.6150435930907727ms
Select for-of benchmark: 1.549559241372347ms
=== deno ===
Map benchmark: 1.0782ms
Map for-of benchmark: 0.2183ms
Select benchmarks: 1.666ms
Select for-of benchmark: 1.6053ms
```

Not as fast as I'd like, but the other way I was considering is actually slower than generators, which is really bad lol.
The reason the normal select benchmark is so slow compared to the select for-of from before is because `Array.from` is actually really slow for no real reason, so the rework uses a standard for-of loop now, as it's much faster.

The benchmarks only show the `select` function working, however it shows a 50-55% speed increase (ignoring the `toArray` call), which is quite significant.